move async worker to separate file
