+++ adapters/repos/db/repo.go
@@ -26,7 +26,6 @@ import (
 	"github.com/weaviate/weaviate/adapters/repos/db/indexcheckpoint"
 	"github.com/weaviate/weaviate/entities/replication"
 	"github.com/weaviate/weaviate/entities/schema"
-	"github.com/weaviate/weaviate/entities/storobj"
 	"github.com/weaviate/weaviate/usecases/config"
 	"github.com/weaviate/weaviate/usecases/memwatch"
 	"github.com/weaviate/weaviate/usecases/monitoring"
@@ -323,58 +322,6 @@ func (db *DB) batchWorker(first bool) {
 
 			objectCounter = 0
 			checkTime = time.Now().Add(time.Second)
-		}
-	}
-}
-
-type job struct {
-	object  *storobj.Object
-	status  objectInsertStatus
-	index   int
-	ctx     context.Context
-	batcher *objectsBatcher
-
-	// async only
-	indexer batchIndexer
-	ids     []uint64
-	vectors [][]float32
-	done    func()
-}
-
-func asyncWorker(ch chan job, logger logrus.FieldLogger, retryInterval time.Duration) {
-	for job := range ch {
-		stop := func() bool {
-			defer job.done()
-
-			for {
-				err := job.indexer.AddBatch(job.ctx, job.ids, job.vectors)
-				if err == nil {
-					return false
-				}
-
-				if errors.Is(err, context.Canceled) {
-					logger.WithError(err).Debug("skipping indexing batch due to context cancellation")
-					return true
-				}
-
-				logger.WithError(err).Infof("failed to index vectors, retrying in %s", retryInterval.String())
-
-				t := time.NewTimer(retryInterval)
-				select {
-				case <-job.ctx.Done():
-					// drain the timer
-					if !t.Stop() {
-						<-t.C
-					}
-
-					return true
-				case <-t.C:
-				}
-			}
-		}()
-
-		if stop {
-			return
 		}
 	}
 }
